Clean up a bit. Still more to do...
